Added a toggle of CursorLine on SpellCheck Toggle
Cursorline enables the line your selecting to change the background to a
different color (this is very handy!), but when spell check is enabled,
it changes the background color to another color, which is then
overridden by cursorline. This prevents you from seeing if a word on the
line your selecting is spelled incorrectly. Which is anoying at best.

This toggles CursorLine off when you turn Spell Check on/off accordingly